In progress: added 'addDefaults' middleware. Action creators still need fixed
